+++ src/renderer/src/components/ModelLibrary/providers/index.ts
@@ -15,6 +15,7 @@ export * from './KimiPlanParameters';
 export * from './LitellmParameters';
 export * from './MinimaxParameters';
 export * from './MistralParameters';
+export * from './NeuralwattParameters';
 export * from './LmStudioParameters';
 export * from './OllamaParameters';
 export * from './OpenAiParameters';