+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - added /subtask command to create and switch to a subtask
 - added file watch mode setting to configure file change detection method
 - properly focusing search input in task sidebar
+- added direct support for Neuralwatt provider
 
 ## [0.66.0]